feat: extract paths from router; rename and move LocCreation.
logion-network/logion-internal#1197
