+++ src/loc/TransactionLocRequestForm.tsx
@@ -10,7 +10,7 @@ import { DraftRequest, LocsState } from "@logion/client";
 import { useUserContext } from "../wallet-user/UserContext";
 import ButtonGroup from "../common/ButtonGroup";
 import Button from "../common/Button";
-import { locDetailsPath } from "../wallet-user/UserRouter";
+import { locDetailsPath } from "../wallet-user/UserPaths";
 import { useNavigate } from "react-router";
 
 export interface FormValues {
